[Feature/#12] 로그아웃 실패해도 로그인 페이지로 이동되는 오류 수정

+++ api/account/fetchers/post-logout.ts
@@ -1,5 +1,7 @@
 import { accountClient, setAuthorizationHeader } from "@/api/client";
+import { resetToLogin } from "@/navigation/navigationRef";
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { Alert } from "react-native";
 
 export const postLogout = async () => {
   try {
@@ -7,18 +9,21 @@ export const postLogout = async () => {
     if(accessToken){
         setAuthorizationHeader(accessToken);
     }
-    const response = await accountClient.post("/api/logout");
-    console.log(response.data);
+    const res = await accountClient.post("/api/logout");
+    if(res.data.success){
+      await AsyncStorage.clear();
+      resetToLogin();
+    }else{
+      Alert.alert("로그아웃 실패", res.data.message || "서버 문제로 로그아웃 실패");
+    }
 
-    await AsyncStorage.clear();
-    
-    return response.data;
+    return res.data;
+
   } catch (error: any) {
     if (error.response?.data) {
-      console.log("Server response:", error.response.data);
+        Alert.alert("로그아웃 실패", error.response.data.message || "로그아웃 실패");
       return error.response.data;
     }
-    console.log(error);
     throw error;
   }
 };
